SLIDER-630. slider-agent unit tests fail on debian and suse and windows

+++ slider-agent/src/main/python/agent/main.py
@@ -172,7 +172,8 @@ def stop_agent():
     if pid == -1:
       print ("Agent process is not running")
     else:
-      os.kill(pid, signal.SIGKILL)
+      if not IS_WINDOWS:
+        os.kill(pid, signal.SIGKILL)
     os._exit(1)
